refactor: ロジックをカスタムフックとユーティリティ関数に分離
